test phantom error on eject in travis-ci
ExitError: Volume failed to eject (note this error is being sent to
STDOUT despite the process erroring, and hence our new remapping from
previous commit is allowing us to see what is happening).

I originally thought this might be an issue with trying to eject the
disk too soon after mounting that only surfaced in a noisy environment
like travis. However, to mitigate this commit builds a fairly large
pause in to ensure the disk is mounted and ready before dismounting. but
that pause it only on the Destroy() test, so if that's the culprit, we
should see 1/2 tests failing intermittently, rather than 2/2 always.